requirements.txt
Django==4.2
mysqlclient==2.1.1
djangorestframework==3.14.0
python-dotenv==1.0.0
django-cors-headers==4.1.0
pytest==7.4.0
pytest-django==4.5.2
black==23.3.0
flake8==6.0.0
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
python-decouple==3.8